% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Generated example data}
\format{
A list with three elements namely X, Y, Z. Each element has 300 rows.
X has 2 columns, Y has 2 columns and Z has 7 columns.
}
\usage{
data
}
\description{
A generated data set containing three sets of variables: X, Y, Z. The
canonical correlation between X and Y depends on some of the Z variables. The
sample size is 300. Z1-Z5 are the important variables for the varying
correlation between X and Y. Z6-Z7 are the noise variables.
}
\examples{
\donttest{
## load generated example data
data(data, package = "RFCCA")
}
}
\keyword{datasets}
